+++ CHANGELOG.md
@@ -10,9 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Removed unused property "cv" from dataset mock details.
+- Display error alert to user upon login failure
 
 ## [1.0.0]
 
-### Add
+### Added
 
 - Added changelog file